use size_t

+++ teep-broker-app/teec-keystone.cpp
@@ -221,12 +221,12 @@ invoke_command_t ocall_pull_invoke_command()
     return queue.pull_invoke_command();
 }
 
-param_buffer_t ocall_read_invoke_param(int index, unsigned int offset)
+param_buffer_t ocall_read_invoke_param(int index, size_t offset)
 {
     return queue.read_invoke_param(index, offset);
 }
 
-void ocall_write_invoke_param(int index, unsigned int offset, unsigned int size, const char *buf)
+void ocall_write_invoke_param(int index, size_t offset, size_t size, const char *buf)
 {
     //lwsl_hexdump_notice(buf, size);
     queue.write_invoke_param(index, offset, size, buf);
